tests/requirements-test-base.txt
pytest==5.3.2
pytest-cov==2.8.1
pytest-django==3.8.0
djangorestframework==3.11.0
sentry-sdk==0.14.2
